+++ SV22T1020362.Admin/Controllers/CategoryController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using SV22T1020362.BusinessLayers;
 using SV22T1020362.Models.Catalog;
 using SV22T1020362.Models.Common;
@@ -7,7 +8,9 @@ namespace SV22T1020362.Admin.Controllers
 {
     /// <summary>
     /// Các chức năng liên quan đến loại hàng
+    /// Chỉ Administrator và DataManager mới có quyền truy cập.
     /// </summary>
+    [Authorize(Roles = $"{WebUserRoles.Administrator},{WebUserRoles.DataManager}")]
     public class CategoryController : Controller
     {
         public const int PAGESIZE = 10;