+++ SV22T1020362.Admin/Controllers/SupplierController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using SV22T1020362.BusinessLayers;
 using SV22T1020362.Models.Common;
 using SV22T1020362.Models.Partner;
@@ -8,6 +9,7 @@ namespace SV22T1020362.Admin.Controllers
     /// <summary>
     /// Các chức năng liên quan đến nhà cung cấp
     /// </summary>
+    [Authorize(Roles = $"{WebUserRoles.Administrator},{WebUserRoles.DataManager}")]
     public class SupplierController : Controller
     {
         public const string SEARCH_SUPPLIER = "SearchSupplier";